fix range dim axis method

+++ +nix/RangeDimension.m
@@ -40,7 +40,7 @@ classdef RangeDimension < nix.Entity
             end
             
             func_name = strcat(obj.alias, '::axis');
-            axis = nix_mx(func_name, obj.nix_handle, uint64(count), uint64(startIndex));
+            axis = nix_mx(func_name, obj.nix_handle, count, startIndex);
         end
 
     end
